refactor: Order와 OrderTable 도메인 합치기
