+++ src/main/java/kitchenpos/order/application/request/NumberOfGuestsRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.application.request;
+package kitchenpos.order.application.request;
 
 public class NumberOfGuestsRequest {
     private int numberOfGuests;
