+++ src/main/java/kitchenpos/order/application/request/OrderTableRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.application.request;
+package kitchenpos.order.application.request;
 
 public class OrderTableRequest {
     private final Integer numberOfGuests;
