+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -3,9 +3,9 @@ package kitchenpos.menu.application;
 import kitchenpos.menu.Menu;
 import kitchenpos.menu.MenuName;
 import kitchenpos.menu.MenuPrice;
-import kitchenpos.menu.SaveMenuProductsEvent;
 import kitchenpos.menu.application.request.MenuRequest;
 import kitchenpos.menugroup.application.MenuGroupRepository;
+import kitchenpos.menuproduct.SaveMenuProductsEvent;
 import kitchenpos.menuproduct.application.MenuProductRepository;
 import kitchenpos.menuproduct.application.request.MenuProductRequest;
 import kitchenpos.product.Product;
