+++ src/test/java/kitchenpos/ordertable/NumberOfGuestsTest.java
@@ -1,5 +1,6 @@
 package kitchenpos.ordertable;
 
+import kitchenpos.order.NumberOfGuests;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
